+++ lib/koolbase_flutter.dart
@@ -8,4 +8,6 @@ export 'src/storage/storage_client.dart';
 export 'src/database/database_client.dart';
 export 'src/database/database_models.dart';
 export 'src/database/database_query.dart';
+export 'src/realtime/realtime_client.dart';
+export 'src/realtime/realtime_models.dart';
 export 'src/storage/storage_models.dart';